requirements.txt
allure-pytest==2.9.45
pytest==7.1.1
selenium==4.1.3
pytest-xdist==2.5.0
